refactor python langgraph
