+++ mock-langgraph-server/main.py
@@ -3,30 +3,12 @@
 from dotenv import load_dotenv
 load_dotenv()
 
-
-import os
-import json
-import uuid
-import time
-
-
 from typing import Annotated
-from pydantic import BaseModel
-from langchain_core.load import dumps
-from fastapi.responses import StreamingResponse
 from fastapi.middleware.cors import CORSMiddleware
-from fastapi import FastAPI, Depends, Header, HTTPException
-from langchain_core.messages import (
-    AIMessageChunk,
-    AIMessage,
-    ToolMessage,
-    HumanMessage,
-)
+from fastapi import FastAPI, Depends
 
 # Utils and modules
-from graph import graph
 from auth import get_authenticated_user
-from database import db_manager
 
 app = FastAPI(title="LangGraph Azure Inference API", version="1.0.0")
 
@@ -39,86 +21,6 @@ app.add_middleware(
     allow_headers=["*"],  # Allow all headers
 )
 
-# Utils
-def generate_uuid():
-    """Generate a unique UUID string."""
-    return str(uuid.uuid4())
-
-# Inference function
-async def generate_stream(input_message: str, conversation_id: str):
-    # Generate unique message ID
-    message_id = str(uuid.uuid4())
-    
-    # Send StartStep (f:) - Start of message processing
-    yield f"f:{json.dumps({'messageId': message_id})}\n"
-    
-    tool_calls = {}
-    tool_calls_by_idx = {}
-    accumulated_text = ""
-    token_count = 0
-
-    try:
-        async for msg, metadata in graph.astream(
-            {"messages": input_message},
-            config={"configurable": {"thread_id": conversation_id}},
-            stream_mode="messages",
-        ):
-            
-            if isinstance(msg, ToolMessage):
-                # Handle tool results - ToolCallResult (a:)
-                tool_call_id = msg.tool_call_id
-                yield f"a:{json.dumps({'toolCallId': tool_call_id, 'result': msg.content})}\n"
-
-            elif isinstance(msg, AIMessageChunk) or isinstance(msg, AIMessage):
-                # Handle text content - TextDelta (0:)
-                if msg.content:
-                    # Send text delta - properly escape the content
-                    content = str(msg.content)
-                    yield f"0:{json.dumps(content)}\n"
-                    accumulated_text += content
-                    token_count += len(content.split())
-
-                # Handle tool calls
-                if hasattr(msg, 'tool_calls') and msg.tool_calls:
-                    for tool_call in msg.tool_calls:
-                        tool_call_id = tool_call.get('id', str(uuid.uuid4()))
-                        tool_name = tool_call.get('name', '')
-
-                        if tool_name == "":
-                            continue
-
-                        tool_calls_by_idx[len(tool_calls_by_idx)] = tool_call_id
-                        tool_calls[tool_call_id] = {"name": tool_name, "args": ""}
-                        
-                        # Send StartToolCall (b:)
-                        yield f"b:{json.dumps({'toolCallId': tool_call_id, 'toolName': tool_name})}\n"
-                        
-                        # # Send ToolCall (9:) with complete args
-                        # yield f"9:{json.dumps({'toolCallId': tool_call_id, 'toolName': tool_name, 'args': tool_args})}\n"
-                
-                # Handle streaming tool call chunks
-                if hasattr(msg, 'tool_call_chunks') and msg.tool_call_chunks:
-                    for chunk in msg.tool_call_chunks:
-                        tool_name = chunk.get("name", "")
-                        args_chunk = chunk.get("args", "")
-                        chunk_index = chunk.get("index", 0)
-                        tool_call_id = tool_calls_by_idx.get(chunk_index, -1)
-
-                        # Accumulate args and send ToolCallArgsTextDelta (c:)
-                        if tool_call_id != -1 and args_chunk:
-                            tool_calls[tool_call_id]["args"] += args_chunk
-                            yield f"c:{json.dumps({'toolCallId': tool_call_id, 'argsTextDelta': args_chunk})}\n"
-                        
-
-        # Send FinishMessage (d:) with usage stats
-        yield f"d:{json.dumps({'finishReason': 'stop', 'usage': {'promptTokens': token_count, 'completionTokens': token_count}})}\n"
-        
-    except Exception as e:
-        # Send Error (3:)
-        error_message = str(e)
-        yield f"3:{json.dumps(error_message)}\n"
-
-
 @app.get("/")
 async def root(username: Annotated[str, Depends(get_authenticated_user)]):
     """Root endpoint."""
@@ -126,179 +28,13 @@ async def root(username: Annotated[str, Depends(get_authenticated_user)]):
 
 
 @app.get("/health")
-async def health(username: Annotated[str, Depends(get_authenticated_user)]):
+async def health():
     """Health check endpoint."""
-    return {"status": "healthy", "authenticated_user": username}
-
-class ChatRequest(BaseModel):
-    messages: list
-
-@app.post("/chat")
-async def chat_completions(request: ChatRequest, _: Annotated[str, Depends(get_authenticated_user)], userid:  Annotated[str | None, Header()] = None):
-    """Chat completions endpoint."""
-
-    if not userid:
-        return {"error": "Missing userid header"}
-
-    conversation_id = generate_uuid()
-    input_message = request.messages[-1] if request.messages else ""
-
-    # Add user and the conversation id to the database
-    db_manager.create_conversation(conversation_id, userid)
-
-    return StreamingResponse(
-        generate_stream(input_message, conversation_id),
-        media_type="text/event-stream",
-        headers={
-            "Cache-Control": "no-cache",
-            "Content-Type": "text/plain; charset=utf-8",
-            "Connection": "keep-alive",
-            "x-vercel-ai-data-stream": "v1",
-            "x-vercel-ai-ui-message-stream": "v1"
-        }
-    )
+    return {"status": "healthy"}
 
 
-@app.get("/last-conversation-id")
-def get_last_conversation_id(_: Annotated[str, Depends(get_authenticated_user)], userid:  Annotated[str | None, Header()] = None):
-    """Get last conversation ID endpoint."""
-    if not userid:
-        return {"error": "Missing userid header"}
-    
-    # Fetch the last conversation ID for the user from the database
-    last_conversation_id = db_manager.get_last_conversation_id(userid)
-
-    return {
-        "userId": userid,
-        "lastConversationId": last_conversation_id
-    }
-
-@app.get("/conversations")
-def get_conversations(_: Annotated[str, Depends(get_authenticated_user)], userid:  Annotated[str | None, Header()] = None):
-    """Get conversations endpoint."""
-    if not userid:
-        return {"error": "Missing userid header"}
-
-    # Fetch list of conversations for the user from the database
-    conversations = db_manager.get_user_conversations(userid)
-    
-    # Convert to the expected API response format
-    response = []
-    for conv in conversations:
-        # Get the first message from the conversation to use as title
-        # For now, we'll use a default title since we don't store message content in metadata
-        # In a real implementation, you might want to fetch the first message from LangGraph state
-        conv_graph_val = graph.get_state(config={"configurable": {"thread_id": conv.id}}).values
-        conv_graph_messages = conv_graph_val.get("messages", []) if conv_graph_val else []
-        title = f"Conversations {conv.id[:8]}..."
-
-        if conv_graph_messages:
-            first_message = conv_graph_messages[0]
-            content = first_message.content
-
-            if isinstance(content, list) and len(content) > 0 and content[0]['type'] == 'text':
-                title = content[0]['text']
-            elif type(content) is str:
-                title = content
-
-        response.append({
-            "id": conv.id,
-            "title": title,
-            "created_at": conv.created_at,
-            "is_pinned": conv.is_pinned
-        })
-    
-    return response
-
-@app.get("/conversations/{conversation_id}")
-def get_chat_history(_: Annotated[str, Depends(get_authenticated_user)], userid:  Annotated[str | None, Header()] = None, conversation_id: str = ""):
-    """Get chat history for a conversation."""
-    if not userid:
-        return {"error": "Missing userid header"}
-    
-    # Check if the conversation exists and belongs to the user
-    if not db_manager.conversation_exists(conversation_id, userid):
-        raise HTTPException(status_code=404, detail="Conversation not found")
-    
-    # Fetch chat history for the conversation from LangGraph state
-    try:
-        # Get the conversation state from the checkpointer
-        states_generator = graph.get_state_history(config={"configurable": {"thread_id": conversation_id}})
-        states = list(states_generator)
-
-        json_dumps = dumps(states)
-        
-        return json.loads(json_dumps)
-        
-    except Exception as e:
-        raise HTTPException(status_code=500, detail=f"Failed to fetch chat history: {str(e)}")
-
-@app.post("/conversations/{conversation_id}/chat")
-def chat_conversation(_: Annotated[str, Depends(get_authenticated_user)], userid: Annotated[str | None, Header()] = None, conversation_id: str = "", request: ChatRequest = None):
-    """Chat in a specific conversation."""
-
-    if not userid:
-        return {"error": "Missing userid header"}
-    
-    if not request:
-        return {"error": "Missing request body"}
-
-    input_message = request.messages[-1] if request.messages else ""
-
-    # Check if the conversation exists and belongs to the user
-    if not db_manager.conversation_exists(conversation_id, userid):
-        raise HTTPException(status_code=404, detail="Conversation not found")
-
-    return StreamingResponse(
-        generate_stream(input_message, conversation_id),
-        media_type="text/event-stream",
-        headers={
-            "Cache-Control": "no-cache",
-            "Content-Type": "text/plain; charset=utf-8",
-            "Connection": "keep-alive",
-            "x-vercel-ai-data-stream": "v1",
-            "x-vercel-ai-ui-message-stream": "v1"
-        }
-    )
-@app.delete("/conversations/{conversation_id}")
-def delete_conversation(_: Annotated[str, Depends(get_authenticated_user)], userid: Annotated[str | None, Header()] = None, conversation_id: str = ""):
-    """Delete a conversation."""
-
-    if not userid:
-        return {"error": "Missing userid header"}
-
-    # Delete the conversation from the database
-    deleted = db_manager.delete_conversation(conversation_id, userid)
-    
-    if not deleted:
-        raise HTTPException(status_code=404, detail="Conversation not found")
-    
-    return {"message": "Conversation deleted successfully"}
-
-@app.post("/conversations/{conversation_id}/pin")
-def pin_conversation(_: Annotated[str, Depends(get_authenticated_user)], userid: Annotated[str | None, Header()] = None, conversation_id: str = ""):
-    """Pin or unpin a conversation."""
-
-    if not userid:
-        return {"error": "Missing userid header"}
-    
-    existing_data = db_manager.get_conversation(conversation_id, userid)
-    if not existing_data:
-        raise HTTPException(status_code=404, detail="Conversation not found")
-
-    # Pin or unpin the conversation in the database
-    updated = db_manager.pin_conversation(conversation_id, userid, not existing_data.is_pinned)
-    
-    if not updated:
-        raise HTTPException(status_code=404, detail="Conversation not found")
-    
-    action = "pinned" if updated else "unpinned"
-    return {"message": f"Conversation {action} successfully"}
-
-if __name__ == "__main__":
-    import uvicorn
-    
-    host = os.getenv("HOST", "0.0.0.0")
-    port = int(os.getenv("PORT", 8000))
-    
-    uvicorn.run(app, host=host, port=port)
+# Add external routers
+from routes.chat_conversation import chat_conversation_route
+app.include_router(
+    chat_conversation_route
+)
