+++ mock-langgraph-server/utils/uuid.py
@@ -1,0 +1,5 @@
+import uuid
+
+def generate_uuid():
+    """Generate a unique UUID string."""
+    return str(uuid.uuid4())
